+++ src/Test/Plutip/Contract.hs
@@ -181,6 +181,7 @@ import PlutusPrelude (render)
 import PlutusTx.These (These (That, These, This))
 import Prettyprinter (Doc, Pretty (pretty), vcat, (<+>))
 import Test.Plutip.Contract.Init (
+  ada,
   initAda,
   initAdaAssertValue,
   initAdaAssertValueWith,
@@ -235,7 +236,6 @@ import Test.Plutip.Options (
   TraceOption (ShowBudgets, ShowTrace, ShowTraceButOnlyContext),
  )
 import Test.Plutip.Predicate (Predicate, noBudgetsMessage, pTag)
-import Test.Plutip.Tools.Cluster (ada)
 import Test.Plutip.Tools.Format (fmtTxBudgets)
 import Test.Tasty (testGroup, withResource)
 import Test.Tasty.HUnit (assertFailure, testCase)